+++ base/android/java/src/org/chromium/base/BaseChromiumApplication.java
@@ -4,11 +4,8 @@
 
 package org.chromium.base;
 
-import android.app.Activity;
 import android.app.Application;
 import android.content.Context;
-import android.os.Bundle;
-import android.view.Window;
 
 import org.chromium.base.multidex.ChromiumMultiDexInstaller;
 
@@ -16,14 +13,7 @@ import org.chromium.base.multidex.ChromiumMultiDexInstaller;
  * Basic application functionality that should be shared among all browser applications.
  */
 public class BaseChromiumApplication extends Application {
-
     private static final String TAG = "cr.base";
-    private static final String TOOLBAR_CALLBACK_INTERNAL_WRAPPER_CLASS =
-            "android.support.v7.internal.app.ToolbarActionBar$ToolbarCallbackWrapper";
-    // In builds using the --use_unpublished_apis flag, the ToolbarActionBar class name does not
-    // include the "internal" package.
-    private static final String TOOLBAR_CALLBACK_WRAPPER_CLASS =
-            "android.support.v7.app.ToolbarActionBar$ToolbarCallbackWrapper";
     private final boolean mShouldInitializeApplicationStatusTracking;
 
     public BaseChromiumApplication() {
@@ -40,42 +30,9 @@ public class BaseChromiumApplication extends Application {
         ChromiumMultiDexInstaller.install(this);
     }
 
-    /**
-     * Interface to be implemented by listeners for window focus events.
-     */
-    public interface WindowFocusChangedListener {
-        /**
-         * Called when the window focus changes for {@code activity}.
-         * @param activity The {@link Activity} that has a window focus changed event.
-         * @param hasFocus Whether or not {@code activity} gained or lost focus.
-         */
-        public void onWindowFocusChanged(Activity activity, boolean hasFocus);
-    }
-
-    private ObserverList<WindowFocusChangedListener> mWindowFocusListeners =
-            new ObserverList<WindowFocusChangedListener>();
-
-    @Override
     public void onCreate() {
         super.onCreate();
-
-        if (mShouldInitializeApplicationStatusTracking) startTrackingApplicationStatus();
-    }
-
-    /**
-     * Registers a listener to receive window focus updates on activities in this application.
-     * @param listener Listener to receive window focus events.
-     */
-    public void registerWindowFocusChangedListener(WindowFocusChangedListener listener) {
-        mWindowFocusListeners.addObserver(listener);
-    }
-
-    /**
-     * Unregisters a listener from receiving window focus updates on activities in this application.
-     * @param listener Listener that doesn't want to receive window focus events.
-     */
-    public void unregisterWindowFocusChangedListener(WindowFocusChangedListener listener) {
-        mWindowFocusListeners.removeObserver(listener);
+        ApplicationStatusManager.init(this);
     }
 
     /** Initializes the {@link CommandLine}. */
@@ -88,79 +45,5 @@ public class BaseChromiumApplication extends Application {
     @VisibleForTesting
     public static void initCommandLine(Context context) {
         ((BaseChromiumApplication) context.getApplicationContext()).initCommandLine();
-    }
-
-    /** Register hooks and listeners to start tracking the application status. */
-    private void startTrackingApplicationStatus() {
-        ApplicationStatus.initialize(this);
-        registerActivityLifecycleCallbacks(new ActivityLifecycleCallbacks() {
-            @Override
-            public void onActivityCreated(final Activity activity, Bundle savedInstanceState) {
-                Window.Callback callback = activity.getWindow().getCallback();
-                activity.getWindow().setCallback(new WindowCallbackWrapper(callback) {
-                    @Override
-                    public void onWindowFocusChanged(boolean hasFocus) {
-                        super.onWindowFocusChanged(hasFocus);
-                        for (WindowFocusChangedListener listener : mWindowFocusListeners) {
-                            listener.onWindowFocusChanged(activity, hasFocus);
-                        }
-                    }
-                });
-            }
-
-            @Override
-            public void onActivityDestroyed(Activity activity) {
-                assert (activity.getWindow().getCallback() instanceof WindowCallbackWrapper
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_WRAPPER_CLASS)
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_INTERNAL_WRAPPER_CLASS));
-            }
-
-            @Override
-            public void onActivityPaused(Activity activity) {
-                assert (activity.getWindow().getCallback() instanceof WindowCallbackWrapper
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_WRAPPER_CLASS)
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_INTERNAL_WRAPPER_CLASS));
-            }
-
-            @Override
-            public void onActivityResumed(Activity activity) {
-                assert (activity.getWindow().getCallback() instanceof WindowCallbackWrapper
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_WRAPPER_CLASS)
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_INTERNAL_WRAPPER_CLASS));
-            }
-
-            @Override
-            public void onActivitySaveInstanceState(Activity activity, Bundle outState) {
-                assert (activity.getWindow().getCallback() instanceof WindowCallbackWrapper
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_WRAPPER_CLASS)
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_INTERNAL_WRAPPER_CLASS));
-            }
-
-            @Override
-            public void onActivityStarted(Activity activity) {
-                assert (activity.getWindow().getCallback() instanceof WindowCallbackWrapper
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_WRAPPER_CLASS)
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_INTERNAL_WRAPPER_CLASS));
-            }
-
-            @Override
-            public void onActivityStopped(Activity activity) {
-                assert (activity.getWindow().getCallback() instanceof WindowCallbackWrapper
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_WRAPPER_CLASS)
-                        || activity.getWindow().getCallback().getClass().getName().equals(
-                                TOOLBAR_CALLBACK_INTERNAL_WRAPPER_CLASS));
-            }
-        });
     }
 }
